create class CompressedArray
The InnerCompressedArray class represents the array that has been compressed
from a matrix (2D array) into a linear array that excludes elements from the
diagonal and any elements above or to the right of the diagonal. Only
elements below or to the left of the matrix diagonal must be included in the
CompressedArray.