Create constructor of the class
1. Takes in a String representing the file to load (i.e. "cities1.txt")
and a boolean (showMap) that indicates whether or not the map GUI should
be displayed.
2. Initialize cityArray with 3 cells.
3. Create an object of MyFileReader or use your own code to read in a
text file, and load in the file with the given name.
4. Read in each line from the file and create a City object containing
the city name, and the x and y values from the file (look at the text
file to see this format). Add the city object to the cityArray and
expand the capacity if needed.
5. If the boolean (showMap) is true, then create a Map object and call
addCity() on the Map object for each city in the cityArray. This will
add the marker icons to the map for each city.