create compareDistances method
the method will Create a 2D double array (i.e. double[][]) with a size of N by N,
where N is the number of cities in cityArray. Loop through every combination
of pairs of cities and call distBetweenCities() for each pair.
Save this result into the double[][] array in the appropriate cell.